Update Assignment8.c
I could not understand the logic of your program. It may be right. frankly speaking, I did not read each statement. However, if a program can be written in the simplest way, it must be. I am attaching here an alternative code. It may be useful for you. 
By the way, your attempt to write such a code is really wonderful.