Update Assignment6.c
In this, you are supposed to populate the array first say of 10 integers having array size greater than that. And then you are attempting to insert an integer at a specific location of an array. So you will have to shift subsequent numbers by one location and print the final array